refactor(cli): centralize Subtensor endpoint resolution
