+++ gittensor/cli/miner_commands/helpers.py
@@ -10,7 +10,7 @@ from pathlib import Path
 import click
 from rich.console import Console
 
-from gittensor.constants import NETWORK_MAP
+from gittensor.cli.endpoint_resolution import resolve_ws_endpoint
 
 console = Console()
 
@@ -41,18 +41,8 @@ def _load_config_value(key: str):
 
 
 def _resolve_endpoint(network: str | None, rpc_url: str | None) -> str:
-    """Resolve the subtensor endpoint from CLI args or config."""
-    if rpc_url:
-        return rpc_url
-    if network:
-        return NETWORK_MAP.get(network.lower(), network)
-    config_network = _load_config_value('network')
-    config_endpoint = _load_config_value('ws_endpoint')
-    if config_endpoint:
-        return config_endpoint
-    if config_network:
-        return NETWORK_MAP.get(config_network.lower()) or config_network
-    return NETWORK_MAP['finney']
+    """Resolve the subtensor endpoint (same rules as issue CLI `resolve_network`)."""
+    return resolve_ws_endpoint(network, rpc_url)
 
 
 def _connect_bittensor(wallet_name: str, wallet_hotkey: str, ws_endpoint: str, netuid: int):
